refactor(strings): move string API to types module
