+++ ibis/expr/api.py
@@ -3177,727 +3177,6 @@ _add_methods(ir.BooleanColumn, _boolean_column_methods)
 
 
 # ---------------------------------------------------------------------
-# Binary API
-
-
-def hashbytes(
-    arg: ir.BinaryValue | ir.StringValue,
-    how: Literal['md5', 'sha1', 'sha256', 'sha512'] = 'sha256',
-) -> ir.BinaryValue:
-    """Compute the binary hash value of `arg`.
-
-    Parameters
-    ----------
-    arg
-        Expression to hash
-    how
-        Hash algorithm to use
-
-    Returns
-    -------
-    BinaryValue
-        Binary expression
-    """
-    return ops.HashBytes(arg, how).to_expr()
-
-
-_binary_value_methods = {'hashbytes': hashbytes}
-_add_methods(ir.BinaryValue, _binary_value_methods)
-
-
-# ---------------------------------------------------------------------
-# String API
-
-
-def _string_substr(
-    self: ir.StringValue,
-    start: int | ir.IntegerValue,
-    length: int | ir.IntegerValue | None = None,
-) -> ir.StringValue:
-    """Pull substrings out by position and maximum length.
-
-    Parameters
-    ----------
-    self
-        String expression
-    start
-        First character to start splitting, indices start at 0
-    length
-        Maximum length of each substring. If not supplied, searches the entire
-        string
-
-    Returns
-    -------
-    StringValue
-        Found substring
-    """
-    op = ops.Substring(self, start, length)
-    return op.to_expr()
-
-
-def _string_left(
-    self: ir.StringValue, nchars: int | ir.IntegerValue
-) -> ir.StringValue:
-    """Return the `nchars` left-most characters each string in `arg`.
-
-    Parameters
-    ----------
-    self
-        String expression
-    nchars
-        Maximum number of characters to return
-
-    Returns
-    -------
-    StringValue
-        Characters
-    """
-    return self.substr(0, length=nchars)
-
-
-def _string_right(
-    self: ir.StringValue, nchars: int | ir.IntegerValue
-) -> ir.StringValue:
-    """Return up to `nchars` from the end of each string in `arg`.
-
-    Parameters
-    ----------
-    self
-        String expression
-    nchars
-        Maximum number of characters to return
-
-    Returns
-    -------
-    StringValue
-        Characters
-    """
-    return ops.StrRight(self, nchars).to_expr()
-
-
-def repeat(self: ir.StringValue, n: int | ir.IntegerValue) -> ir.StringValue:
-    """Repeat the string `self` `n` times.
-
-    Parameters
-    ----------
-    self
-        String to repeat
-    n
-        Number of repetitions
-
-    Returns
-    -------
-    StringValue
-        Repeated string
-    """
-    return ops.Repeat(self, n).to_expr()
-
-
-def _translate(
-    self: ir.StringValue, from_str: ir.StringValue, to_str: ir.StringValue
-) -> ir.StringValue:
-    """Replace `from_str` characters in `self` characters in `to_str`.
-
-    To avoid unexpected behavior, `from_str` should be shorter than `to_str`.
-
-    Parameters
-    ----------
-    self
-        String expression
-    from_str
-        Characters in `arg` to replace
-    to_str
-        Characters to use for replacement
-
-    Examples
-    --------
-    >>> import ibis
-    >>> table = ibis.table([('string_col', 'string')])
-    >>> expr = table.string_col.translate('a', 'b')
-    >>> expr = table.string_col.translate('a', 'bc')
-
-    Returns
-    -------
-    StringValue
-        Translated string
-    """
-    return ops.Translate(self, from_str, to_str).to_expr()
-
-
-def _string_find(
-    self: ir.StringValue,
-    substr: str | ir.StringValue,
-    start: int | ir.IntegerValue | None = None,
-    end: int | ir.IntegerValue | None = None,
-) -> ir.IntegerValue:
-    """Return the position of the first occurence of substring.
-
-    Parameters
-    ----------
-    self
-        String expression
-    substr
-        Substring to search for
-    start
-        Zero based index of where to start the search
-    end
-        Zero based index of where to stop the search. Currently not
-        implemented.
-
-    Returns
-    -------
-    IntegerValue
-        Position of `substr` in `arg` starting from `start`
-    """
-    if end is not None:
-        raise NotImplementedError
-    return ops.StringFind(self, substr, start, end).to_expr()
-
-
-def _lpad(
-    self: ir.StringValue,
-    length: int | ir.IntegerValue,
-    pad: str | ir.StringValue = ' ',
-) -> ir.StringValue:
-    """Pad `arg` by truncating on the right or padding on the left.
-
-    Parameters
-    ----------
-    self
-        String to pad
-    length : int
-        Length of output string
-    pad
-        Pad character
-
-    Examples
-    --------
-    >>> import ibis
-    >>> table = ibis.table([('strings', 'string')])
-    >>> expr = table.strings.lpad(5, '-')
-    >>> expr = ibis.literal('a').lpad(5, '-')  # 'a' becomes '----a'
-    >>> expr = ibis.literal('abcdefg').lpad(5, '-')  # 'abcdefg' becomes 'abcde'  # noqa: E501
-
-    Returns
-    -------
-    StringValue
-        Padded string
-    """
-    return ops.LPad(self, length, pad).to_expr()
-
-
-def _rpad(
-    self: ir.StringValue,
-    length: int | ir.IntegerValue,
-    pad: str | ir.StringValue = ' ',
-) -> ir.StringValue:
-    """Pad `self` by truncating or padding on the right.
-
-    Parameters
-    ----------
-    self
-        String to pad
-    length : int
-        Length of output string
-    pad
-        Pad character
-
-    Examples
-    --------
-    >>> import ibis
-    >>> table = ibis.table([('string_col', 'string')])
-    >>> expr = table.string_col.rpad(5, '-')
-    >>> expr = ibis.literal('a').rpad(5, '-')  # 'a' becomes 'a----'
-    >>> expr = ibis.literal('abcdefg').rpad(5, '-')  # 'abcdefg' becomes 'abcde'  # noqa: E501
-
-    Returns
-    -------
-    StringValue
-        Padded string
-    """
-    return ops.RPad(self, length, pad).to_expr()
-
-
-def _find_in_set(
-    self: ir.StringValue, str_list: Sequence[str]
-) -> ir.IntegerValue:
-    """Find the first occurence of `str_list` within a list of strings.
-
-    No string in list can have a comma.
-
-    Parameters
-    ----------
-    self
-        String expression to search
-    str_list
-        Sequence of strings
-
-    Examples
-    --------
-    >>> import ibis
-    >>> table = ibis.table([('strings', 'string')])
-    >>> result = table.strings.find_in_set(['a', 'b'])
-
-    Returns
-    -------
-    IntegerValue
-        Position of `str_list` in `arg`. Returns -1 if `arg` isn't found or if
-        `arg` contains `','`.
-    """
-    return ops.FindInSet(self, str_list).to_expr()
-
-
-def _string_join(
-    self: str | ir.StringValue, strings: Sequence[str]
-) -> ir.StringValue:
-    """Join a list of strings using the `self` as the separator.
-
-    Parameters
-    ----------
-    self
-        String expression
-    strings
-        Strings to join with `arg`
-
-    Examples
-    --------
-    >>> import ibis
-    >>> sep = ibis.literal(',')
-    >>> result = sep.join(['a', 'b', 'c'])
-
-    Returns
-    -------
-    StringValue
-        Joined string
-    """
-    return ops.StringJoin(self, strings).to_expr()
-
-
-def _startswith(
-    self: str | ir.StringValue, start: str | ir.StringValue
-) -> ir.BooleanValue:
-    """Determine whether `self` starts with `end`.
-
-    Parameters
-    ----------
-    self
-        String expression
-    start
-        prefix to check for
-
-    Examples
-    --------
-    >>> import ibis
-    >>> text = ibis.literal('Ibis project')
-    >>> text.startswith('Ibis')
-
-    Returns
-    -------
-    BooleanValue
-        Boolean indicating whether `self` starts with `start`
-    """
-    return ops.StartsWith(self, start).to_expr()
-
-
-def _endswith(
-    self: str | ir.StringValue, end: str | ir.StringValue
-) -> ir.BooleanValue:
-    """Determine if `self` ends with `end`.
-
-    Parameters
-    ----------
-    self
-        String expression
-    end
-        Suffix to check for
-
-    Examples
-    --------
-    >>> import ibis
-    >>> text = ibis.literal('Ibis project')
-    >>> text.endswith('project')
-
-    Returns
-    -------
-    BooleanValue
-        Boolean indicating whether `self` ends with `end`
-    """
-    return ops.EndsWith(self, end).to_expr()
-
-
-def _string_like(
-    self: ir.StringValue,
-    patterns: str | ir.StringValue | Sequence[str | ir.StringValue],
-) -> ir.BooleanValue:
-    """Match `patterns` against `self`, case-sensitive.
-
-    This function is modeled after the SQL `LIKE` directive. Use `%` as a
-    multiple-character wildcard or `_` as a single-character wildcard.
-
-    Use `re_search` or `rlike` for regular expression-based matching.
-
-    Parameters
-    ----------
-    self
-        String expression
-    patterns
-        If `pattern` is a list, then if any pattern matches the input then
-        the corresponding row in the output is `True`.
-
-    Returns
-    -------
-    BooleanValue
-        Column indicating matches
-    """
-    return functools.reduce(
-        operator.or_,
-        (
-            ops.StringSQLLike(self, pattern).to_expr()
-            for pattern in util.promote_list(patterns)
-        ),
-    )
-
-
-def _string_ilike(
-    self: ir.StringValue,
-    patterns: str | ir.StringValue | Sequence[str | ir.StringValue],
-) -> ir.BooleanValue:
-    """Match `patterns` against `self`, case-insensitive.
-
-    This function is modeled after the SQL `ILIKE` directive. Use `%` as a
-    multiple-character wildcard or `_` as a single-character wildcard.
-
-    Use `re_search` or `rlike` for regular expression-based matching.
-
-    Parameters
-    ----------
-    self
-        String expression
-    patterns
-        If `pattern` is a list, then if any pattern matches the input then
-        the corresponding row in the output is `True`.
-
-    Returns
-    -------
-    BooleanValue
-        Column indicating matches
-    """
-    return functools.reduce(
-        operator.or_,
-        (
-            ops.StringSQLILike(self, pattern).to_expr()
-            for pattern in util.promote_list(patterns)
-        ),
-    )
-
-
-def re_search(
-    arg: str | ir.StringValue, pattern: str | ir.StringValue
-) -> ir.BooleanValue:
-    """Return whether the values in `arg` match `pattern`.
-
-    Returns `True` if the regex matches a string and `False` otherwise.
-
-    Parameters
-    ----------
-    arg
-        String expression to check for matches
-    pattern
-        Regular expression use for searching
-
-    Returns
-    -------
-    BooleanValue
-        Indicator of matches
-    """
-    return ops.RegexSearch(arg, pattern).to_expr()
-
-
-def regex_extract(
-    arg: str | ir.StringValue,
-    pattern: str | ir.StringValue,
-    index: int | ir.IntegerValue,
-) -> ir.StringValue:
-    """Return the specified match from a regular expression `pattern`.
-
-    Parameters
-    ----------
-    arg
-        String expression
-    pattern
-        Reguar expression string
-    index
-        Zero-based index of match to return
-
-    Returns
-    -------
-    StringValue
-        Extracted match
-    """
-    return ops.RegexExtract(arg, pattern, index).to_expr()
-
-
-def regex_replace(
-    arg: str | ir.StringValue,
-    pattern: str | ir.StringValue,
-    replacement: str | ir.StringValue,
-) -> ir.StringValue:
-    """Replace match found by regex `pattern` with `replacement`.
-
-    `replacement` can also be a regex
-
-    Parameters
-    ----------
-    arg
-        String expression
-    pattern
-        Regular expression string
-    replacement
-        Replacement string; can be a regular expression
-
-    Examples
-    --------
-    >>> import ibis
-    >>> table = ibis.table([('strings', 'string')])
-    >>> result = table.strings.replace('(b+)', r'<\1>')  # 'aaabbbaa' becomes 'aaa<bbb>aaa'  # noqa: E501
-
-    Returns
-    -------
-    StringValue
-        Modified string
-    """
-    return ops.RegexReplace(arg, pattern, replacement).to_expr()
-
-
-def _string_replace(
-    arg: str | ir.StringValue,
-    pattern: ir.StringValue,
-    replacement: ir.StringValue,
-) -> ir.StringValue:
-    """Replace each exact match of `pattern` with `replacement`.
-    string.
-
-    Like Python built-in [`str.replace`][str.replace]
-
-    Parameters
-    ----------
-    arg
-        String expression
-    pattern
-        String pattern
-    replacement
-        String replacement
-
-    Examples
-    --------
-    >>> import ibis
-    >>> table = ibis.table([('strings', 'string')])
-    >>> result = table.strings.replace('aaa', 'foo')  # 'aaabbbaaa' becomes 'foobbbfoo'  # noqa: E501
-
-    Returns
-    -------
-    StringVulae
-        Replaced string
-    """
-    return ops.StringReplace(arg, pattern, replacement).to_expr()
-
-
-def to_timestamp(
-    arg: ir.StringValue, format_str: str, timezone: str | None = None
-) -> ir.TimestampValue:
-    """Parse a string and return a timestamp.
-
-    Parameters
-    ----------
-    format_str
-        Format string in `strptime` format
-    timezone
-        A string indicating the timezone. For example `'America/New_York'`
-
-    Examples
-    --------
-    >>> import ibis
-    >>> date_as_str = ibis.literal('20170206')
-    >>> result = date_as_str.to_timestamp('%Y%m%d')
-
-    Returns
-    -------
-    TimestampValue
-        Parsed timestamp value
-    """
-    return ops.StringToTimestamp(arg, format_str, timezone).to_expr()
-
-
-def parse_url(
-    arg: str | ir.StringValue,
-    extract: Literal[
-        "PROTOCOL",
-        "HOST",
-        "PATH",
-        "REF",
-        "AUTHORITY",
-        "FILE",
-        "USERINFO",
-        "QUERY",
-    ],
-    key: str | None = None,
-) -> ir.StringValue:
-    """Parse a URL and extract its components.
-
-    `key` can be used to extract query values when `extract == 'QUERY'`
-
-    Parameters
-    ----------
-    arg
-        URL to extract from
-    extract : str
-        Component of URL to extract
-    key
-        Query component to extract
-
-    Examples
-    --------
-    >>> url = "https://www.youtube.com/watch?v=kEuEcWfewf8&t=10"
-    >>> parse_url(url, 'QUERY', 'v')  # doctest: +SKIP
-    'kEuEcWfewf8'
-
-    Returns
-    -------
-    StringValue
-        Extracted string value
-    """
-    return ops.ParseURL(arg, extract, key).to_expr()
-
-
-def _string_contains(
-    arg: str | ir.StringValue, substr: str | ir.StringValue
-) -> ir.BooleanValue:
-    """Determine if `arg` contains `substr`.
-
-    Parameters
-    ----------
-    arg
-        String value expression
-    substr
-        Substring to check `arg` for
-
-    Returns
-    -------
-    BooleanValue
-        Boolean value indicating the presence of `substr` in `arg`
-    """
-    return arg.find(substr) >= 0
-
-
-def _string_split(
-    arg: str | ir.StringValue, delimiter: str | ir.StringValue
-) -> ir.ArrayValue:
-    """Split `arg` on `delimiter`.
-
-    Parameters
-    ----------
-    arg
-        String value to split
-    delimiter
-        Value to split by
-
-    Returns
-    -------
-    ArrayValue
-        The string `arg` split by `delimiter`
-    """
-    return ops.StringSplit(arg, delimiter).to_expr()
-
-
-def _string_concat(*args):
-    return ops.StringConcat(args).to_expr()
-
-
-def _string_dunder_contains(arg, substr):
-    raise TypeError('Use val.contains(arg)')
-
-
-def _string_getitem(
-    self: str | ir.StringValue, key: slice | int
-) -> ir.StringValue:
-    if isinstance(key, slice):
-        start, stop, step = key.start, key.stop, key.step
-
-        if step is not None and not isinstance(step, ir.Expr) and step != 1:
-            raise ValueError('Step can only be 1')
-
-        if not isinstance(start, ir.Expr):
-            if start is not None and start < 0:
-                raise ValueError(
-                    'Negative slicing not yet supported, got start value of '
-                    '{:d}'.format(start)
-                )
-            if start is None:
-                start = 0
-
-        if not isinstance(stop, ir.Expr):
-            if stop is not None and stop < 0:
-                raise ValueError(
-                    'Negative slicing not yet supported, got stop value of '
-                    '{:d}'.format(stop)
-                )
-            if stop is None:
-                stop = self.length()
-
-        return self.substr(start, stop - start)
-    elif isinstance(key, int):
-        return self.substr(key, 1)
-    raise NotImplementedError(f'string __getitem__[{type(key).__name__}]')
-
-
-_string_value_methods = {
-    '__getitem__': _string_getitem,
-    'length': _unary_op('length', ops.StringLength),
-    'lower': _unary_op('lower', ops.Lowercase),
-    'upper': _unary_op('upper', ops.Uppercase),
-    'reverse': _unary_op('reverse', ops.Reverse),
-    'ascii_str': _unary_op('ascii', ops.StringAscii),
-    'strip': _unary_op('strip', ops.Strip),
-    'lstrip': _unary_op('lstrip', ops.LStrip),
-    'rstrip': _unary_op('rstrip', ops.RStrip),
-    'capitalize': _unary_op('initcap', ops.Capitalize),
-    'convert_base': convert_base,
-    '__contains__': _string_dunder_contains,
-    'contains': _string_contains,
-    'hashbytes': hashbytes,
-    'like': _string_like,
-    'ilike': _string_ilike,
-    'rlike': re_search,
-    'replace': _string_replace,
-    're_search': re_search,
-    're_extract': regex_extract,
-    're_replace': regex_replace,
-    'to_timestamp': to_timestamp,
-    'parse_url': parse_url,
-    'substr': _string_substr,
-    'left': _string_left,
-    'right': _string_right,
-    'repeat': repeat,
-    'find': _string_find,
-    'translate': _translate,
-    'find_in_set': _find_in_set,
-    'split': _string_split,
-    'join': _string_join,
-    'startswith': _startswith,
-    'endswith': _endswith,
-    'lpad': _lpad,
-    'rpad': _rpad,
-    '__add__': _string_concat,
-    '__radd__': lambda *args: _string_concat(*args[::-1]),
-    '__mul__': _binop_expr('__mul__', ops.Repeat),
-    '__rmul__': _binop_expr('__rmul__', ops.Repeat),
-}
-
-
-_add_methods(ir.StringValue, _string_value_methods)
-
-
-# ---------------------------------------------------------------------
 # Decimal API
